# docs/swagger.yaml
info:
  contact: {}
paths:
...
  /product/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a product
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Delete a product
      tags:
      - Product
    get:
      consumes:
      - application/json
      description: Get a product by ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              description:
                type: string
              id:
                type: integer
              name:
                type: string
              price:
                type: number
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Get a product by ID
      tags:
      - Product
    put:
      consumes:
      - application/json
      description: Update a product
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Product details
        in: body
        name: product
        required: true
        schema:
          properties:
            description:
              type: string
            name:
              type: string
            price:
              type: number
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              description:
                type: string
              id:
                type: integer
              name:
                type: string
              price:
                type: number
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Update a product
      tags:
      - Product
  /product/all:
    get:
      consumes:
      - application/json
      description: Get all products
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              products:
                items:
                  properties:
                    description:
                      type: string
                    id:
                      type: integer
                    name:
                      type: string
                    price:
                      type: number
                  type: object
                type: array
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Get all products
      tags:
      - Product
  /product/create:
    post:
      consumes:
      - application/json
      description: Create a new product
      parameters:
      - description: Product details
        in: body
        name: product
        required: true
        schema:
          properties:
            description:
              type: string
            name:
              type: string
            price:
              type: number
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            properties:
              description:
                type: string
              id:
                type: integer
              name:
                type: string
              price:
                type: number
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Create a new product
      tags:
      - Product
swagger: "2.0"